feat: add limits in CRUD of Billing

+++ server/api/billing/[id].delete.ts
@@ -1,6 +1,6 @@
 import { zh } from 'h3-zod'
 import { eq } from 'drizzle-orm'
-import { billing } from './../../database/schema'
+import { canDeleteBilling } from '~~/shared/abilities/billing'
 
 export default eventHandler(async (event) => {
   const db = useDatabase()
@@ -8,6 +8,7 @@ export default eventHandler(async (event) => {
   const { id } = await zh.useValidatedParams(event, {
     id: zh.intAsString,
   })
+  await authorize(event, canDeleteBilling)
 
   const bill = await db.delete(tables.billing).where(eq(tables.billing.id, id))
